Change size_mat to unsigned long in CumulantK17CompChimStream

+++ src/gpu/VirtualFluids_GPU/Kernel/Kernels/BasicKernels/FluidFlow/Compressible/CumulantK17chimStream/CumulantK17CompChimStream_Device.cuh
@@ -10,7 +10,7 @@ extern "C" __global__ void LB_Kernel_CumulantK17CompChimStream(
 	uint* neighborY,
 	uint* neighborZ,
 	real* distributions,
-	int size_Mat,
+	unsigned long size_Mat,
 	int level,
 	real* forces,
 	real* quadricLimiters,
